refactor(rsc): organize internal plugins
