+++ packages/plugin-rsc/src/plugin.ts
@@ -1,9 +1,8 @@
 import assert from 'node:assert'
 import { createHash } from 'node:crypto'
 import fs from 'node:fs'
-import { createRequire } from 'node:module'
 import path from 'node:path'
-import { fileURLToPath, pathToFileURL } from 'node:url'
+import { pathToFileURL } from 'node:url'
 import { createRequestListener } from '@remix-run/node-fetch-server'
 import * as esModuleLexer from 'es-module-lexer'
 import MagicString from 'magic-string'
@@ -38,9 +37,11 @@ import {
   prepareError,
 } from './vite-utils'
 import { cjsModuleRunnerPlugin } from './plugins/cjs'
-import { evalValue, parseIdQuery } from './plugins/utils'
+import { evalValue, parseIdQuery, resolvePackage } from './plugins/utils'
 import { createDebug } from '@hiogawa/utils'
 import { transformScanBuildStrip } from './plugins/scan'
+import { validateImportPlugin } from './plugins/validate-import'
+import { vitePluginFindSourceMapURL } from './plugins/find-source-map-url'
 
 // state for build orchestration
 let serverReferences: Record<string, string> = {}
@@ -70,12 +71,6 @@ const REACT_SERVER_DOM_NAME = `${PKG_NAME}/vendor/react-server-dom`
 // dev-only wrapper virtual module of rollupOptions.input.index
 const VIRTUAL_ENTRIES = {
   browser: 'virtual:vite-rsc/entry-browser',
-}
-
-const require = createRequire(import.meta.url)
-
-function resolvePackage(name: string) {
-  return pathToFileURL(require.resolve(name)).href
 }
 
 export type RscPluginOptions = {
@@ -1658,102 +1653,6 @@ function collectAssetDepsInner(
 }
 
 //
-// support findSourceMapURL
-// https://github.com/facebook/react/pull/29708
-// https://github.com/facebook/react/pull/30741
-//
-
-export function vitePluginFindSourceMapURL(): Plugin[] {
-  return [
-    {
-      name: 'rsc:findSourceMapURL',
-      apply: 'serve',
-      configureServer(server) {
-        server.middlewares.use(async (req, res, next) => {
-          const url = new URL(req.url!, `http://localhost`)
-          if (url.pathname === '/__vite_rsc_findSourceMapURL') {
-            let filename = url.searchParams.get('filename')!
-            let environmentName = url.searchParams.get('environmentName')!
-            try {
-              const map = await findSourceMapURL(
-                server,
-                filename,
-                environmentName,
-              )
-              res.setHeader('content-type', 'application/json')
-              if (!map) res.statusCode = 404
-              res.end(JSON.stringify(map ?? {}))
-            } catch (e) {
-              next(e)
-            }
-            return
-          }
-          next()
-        })
-      },
-    },
-  ]
-}
-
-export async function findSourceMapURL(
-  server: ViteDevServer,
-  filename: string,
-  environmentName: string,
-): Promise<object | undefined> {
-  // this is likely server external (i.e. outside of Vite processing)
-  if (filename.startsWith('file://')) {
-    filename = fileURLToPath(filename)
-    if (fs.existsSync(filename)) {
-      // line-by-line identity source map
-      const content = fs.readFileSync(filename, 'utf-8')
-      return {
-        version: 3,
-        sources: [filename],
-        sourcesContent: [content],
-        mappings: 'AAAA' + ';AACA'.repeat(content.split('\n').length),
-      }
-    }
-    return
-  }
-
-  // server component stack, replace log, `registerServerReference`, etc...
-  let mod: EnvironmentModuleNode | undefined
-  let map:
-    | NonNullable<EnvironmentModuleNode['transformResult']>['map']
-    | undefined
-  if (environmentName === 'Server') {
-    mod = server.environments.rsc!.moduleGraph.getModuleById(filename)
-    // React extracts stacktrace via resetting `prepareStackTrace` on the server
-    // and let browser devtools handle the mapping.
-    // https://github.com/facebook/react/blob/4a36d3eab7d9bbbfae62699989aa95e5a0297c16/packages/react-server/src/ReactFlightStackConfigV8.js#L15-L20
-    // This means it has additional +2 line offset due to Vite's module runner
-    // function wrapper. We need to correct it just like Vite module runner.
-    // https://github.com/vitejs/vite/blob/d94e7b25564abb81ab7b921d4cd44d0f0d22fec4/packages/vite/src/shared/utils.ts#L58-L69
-    // https://github.com/vitejs/vite/blob/d94e7b25564abb81ab7b921d4cd44d0f0d22fec4/packages/vite/src/node/ssr/fetchModule.ts#L142-L146
-    map = mod?.transformResult?.map
-    if (map && map.mappings) {
-      map = { ...map, mappings: (';;' + map.mappings) as any }
-    }
-  }
-
-  const base = server.config.base.slice(0, -1)
-
-  // `createServerReference(... findSourceMapURL ...)` called on browser
-  if (environmentName === 'Client') {
-    try {
-      const url = new URL(filename).pathname.slice(base.length)
-      mod = server.environments.client.moduleGraph.urlToModuleMap.get(url)
-      map = mod?.transformResult?.map
-    } catch (e) {}
-  }
-
-  if (mod && map) {
-    // fix sources to match Vite's module url on browser
-    return { ...map, sources: [base + mod.url] }
-  }
-}
-
-//
 // css support
 //
 
@@ -2207,49 +2106,6 @@ export function __fix_cloudflare(): Plugin {
       // workaround (fixed in Vite 7) https://github.com/vitejs/vite/pull/20077
       ;(config.environments as any).ssr.resolve.noExternal = true
       ;(config.environments as any).rsc.resolve.noExternal = true
-    },
-  }
-}
-
-// https://github.com/vercel/next.js/blob/90f564d376153fe0b5808eab7b83665ee5e08aaf/packages/next/src/build/webpack-config.ts#L1249-L1280
-// https://github.com/pcattori/vite-env-only/blob/68a0cc8546b9a37c181c0b0a025eb9b62dbedd09/src/deny-imports.ts
-// https://github.com/sveltejs/kit/blob/84298477a014ec471839adf7a4448d91bc7949e4/packages/kit/src/exports/vite/index.js#L513
-function validateImportPlugin(): Plugin {
-  return {
-    name: 'rsc:validate-imports',
-    resolveId: {
-      order: 'pre',
-      async handler(source, importer, options) {
-        // optimizer is not aware of server/client boudnary so skip
-        if ('scan' in options && options.scan) {
-          return
-        }
-
-        // Validate client-only imports in server environments
-        if (source === 'client-only') {
-          if (this.environment.name === 'rsc') {
-            throw new Error(
-              `'client-only' cannot be imported in server build (importer: '${importer ?? 'unknown'}', environment: ${this.environment.name})`,
-            )
-          }
-          return { id: `\0virtual:vite-rsc/empty`, moduleSideEffects: false }
-        }
-        if (source === 'server-only') {
-          if (this.environment.name !== 'rsc') {
-            throw new Error(
-              `'server-only' cannot be imported in client build (importer: '${importer ?? 'unknown'}', environment: ${this.environment.name})`,
-            )
-          }
-          return { id: `\0virtual:vite-rsc/empty`, moduleSideEffects: false }
-        }
-
-        return
-      },
-    },
-    load(id) {
-      if (id.startsWith('\0virtual:vite-rsc/empty')) {
-        return `export {}`
-      }
     },
   }
 }
